refactor : 프로젝트 구조 피드백 반영
