+++ src/main/java/com/pbs/likelion/LikeLionKing/LionStudyApplication.java
@@ -1,0 +1,13 @@
+package com.pbs.likelion.LikeLionKing;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LionStudyApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LionStudyApplication.class, args);
+	}
+
+}
