--- src/main/java/lionStudy/lionStudy/LionStudyApplication.java
@@ -1,13 +1,0 @@
-package lionStudy.lionStudy;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LionStudyApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LionStudyApplication.class, args);
-	}
-
-}
