+++ src/test/java/com/pbs/likelion/LikeLionKing/LionStudyApplicationTests.java
@@ -1,0 +1,13 @@
+package com.pbs.likelion.LikeLionKing;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LionStudyApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
