--- src/test/java/lionStudy/lionStudy/LionStudyApplicationTests.java
@@ -1,13 +1,0 @@
-package lionStudy.lionStudy;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LionStudyApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
